Move CPU tests from unit to integration
