+++ src/lib.rs
@@ -659,160 +659,6 @@ mod tests {
     assert_eq!(spork.drop_history(StatType::Children), None);
   }
 
-  #[test]
-  #[cfg(unix)]
-  fn should_get_linux_process_stats_fib_25() {
-    // intentionally introduce some delays to simulate some weird contention for the clocks among
-    // testing threads in order to hopefully draw out any bugs scoping the results between threads
-    let wait = rand_in_range(100, 400);
-    let expected_cpu = 10_f64;
-
-    let before = utils::now_ms() as u64;
-    let spork = Spork::new().unwrap();
-
-    sleep_ms!(wait);
-    // kick the cpu a bit
-    fib(25);
-
-    let stats = match spork.stats(StatType::Process) {
-      Ok(s) => s,
-      Err(e) => panic!("Stats error {:?}", e)
-    };
-    let _final = utils::now_ms() as u64;
-
-    println!("{:?}", stats);
-    assert!(stats.cpu > expected_cpu);
-    assert!(stats.memory > 0);
-    assert!(stats.duration >= wait);
-    assert!(stats.duration <= _final - before);
-    assert_eq!(stats.cores, 1);
-    assert_eq!(stats.kind, StatType::Process);
-    assert!(stats.uptime >= wait);
-    assert!(stats.uptime <= _final - before);
-    assert!(stats.polled <= _final as i64);
-  }
-
-  #[test]
-  #[cfg(unix)]
-  fn should_get_linux_thread_stats_fib_35() {
-    let wait = rand_in_range(100, 400);
-    let expected_cpu = 10_f64;
-
-    let before = utils::now_ms() as u64;
-    let spork = Spork::new().unwrap();
-
-    sleep_ms!(wait);
-    // kick the cpu a bit
-    fib(35);
-
-    let stats = match spork.stats(StatType::Thread) {
-      Ok(s) => s,
-      Err(e) => panic!("Stats error {:?}", e)
-    };
-    let _final = utils::now_ms() as u64;
-
-    println!("{:?}", stats);
-    assert!(stats.cpu > expected_cpu);
-    assert!(stats.memory > 0);
-    assert!(stats.duration >= wait);
-    assert!(stats.duration <= _final - before);
-    assert_eq!(stats.cores, 1);
-    assert_eq!(stats.kind, StatType::Thread);
-    assert!(stats.uptime >= wait);
-    assert!(stats.uptime <= _final - before);
-    assert!(stats.polled <= _final as i64);
-  }
-
-  #[test]
-  #[cfg(unix)]
-  fn should_get_low_cpu_linux_thread_stats() {
-    let wait = rand_in_range(4000, 6000);
-    let expected_cpu = 1.5_f64;
-
-    let before = utils::now_ms() as u64;
-    let spork = Spork::new().unwrap();
-
-    sleep_ms!(wait);
-
-    let stats = match spork.stats(StatType::Thread) {
-      Ok(s) => s,
-      Err(e) => panic!("Stats error {:?}", e)
-    };
-    let _final = utils::now_ms() as u64;
-
-    println!("{:?}", stats);
-    assert!(stats.cpu < expected_cpu);
-    assert!(stats.memory > 0);
-    assert!(stats.duration >= wait);
-    assert!(stats.duration <= _final - before);
-    assert_eq!(stats.cores, 1);
-    assert_eq!(stats.kind, StatType::Thread);
-    assert!(stats.uptime >= wait);
-    assert!(stats.uptime <= _final - before);
-    assert!(stats.polled <= _final as i64);
-  }
-
-  #[test]
-  #[cfg(unix)]
-  fn should_get_linux_process_stats_with_cpus() {
-    let wait = rand_in_range(100, 400);
-    let expected_cpu = 5_f64;
-
-    let before = utils::now_ms() as u64;
-    let spork = Spork::new().unwrap();
-
-    sleep_ms!(wait);
-    // kick the cpu a bit
-    fib(25);
-
-    let stats = match spork.stats_with_cpus(StatType::Process, Some(spork.num_cores())) {
-      Ok(s) => s,
-      Err(e) => panic!("Stats error {:?}", e)
-    };
-    let _final = utils::now_ms() as u64;
-
-    println!("{:?}", stats);
-    assert!(stats.cpu > expected_cpu);
-    assert!(stats.memory > 0);
-    assert!(stats.duration >= wait);
-    assert!(stats.duration <= _final - before);
-    assert_eq!(stats.cores, spork.num_cores());
-    assert_eq!(stats.kind, StatType::Process);
-    assert!(stats.uptime >= wait);
-    assert!(stats.uptime <= _final - before);
-    assert!(stats.polled <= _final as i64);
-  }
-
-  #[test]
-  #[cfg(unix)]
-  fn should_get_linux_thread_stats_with_cpus() {
-    let wait = rand_in_range(100, 400);
-    let expected_cpu = 5_f64;
-
-    let before = utils::now_ms() as u64;
-    let spork = Spork::new().unwrap();
-
-    sleep_ms!(wait);
-    // kick the cpu a bit
-    fib(35);
-
-    let stats = match spork.stats_with_cpus(StatType::Thread, Some(spork.num_cores())) {
-      Ok(s) => s,
-      Err(e) => panic!("Stats error {:?}", e)
-    };
-    let _final = utils::now_ms() as u64;
-
-    println!("{:?}", stats);
-    assert!(stats.cpu > expected_cpu);
-    assert!(stats.memory > 0);
-    assert!(stats.duration >= wait);
-    assert!(stats.duration <= _final - before);
-    assert_eq!(stats.cores, spork.num_cores());
-    assert_eq!(stats.kind, StatType::Thread);
-    assert!(stats.uptime >= wait);
-    assert!(stats.uptime <= _final - before);
-    assert!(stats.polled <= _final as i64);
-  }
 
   #[test]
   #[cfg(windows)]
